requirements.txt
alembic==1.10.2
Babel==2.12.1
blinker==1.5
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
dominate==2.7.0
elastic-transport==8.4.0
elasticsearch==8.8.0
email-validator==1.3.1
emoji==2.5.1
Flask==2.2.3
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
html5lib==1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langdetect==1.0.9
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.1
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2023.3
regex==2023.6.3
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.6
textile==4.0.2
typing_extensions==4.5.0
urllib3==1.26.15
visitor==0.1.3
webencodings==0.5.1
Werkzeug==2.2.3
WTForms==3.0.1
